Extracted initiateAndSendCatchUpRequest(), initiateAndSendCatchUpResponse() and handleVoteMessage() methods from GrandpaSetState.java to GrandpaMessageHandler.java.
