+++ src/main/java/com/limechain/network/protocol/grandpa/GrandpaEngine.java
@@ -2,7 +2,6 @@ package com.limechain.network.protocol.grandpa;
 
 import com.limechain.exception.scale.ScaleEncodingException;
 import com.limechain.grandpa.GrandpaService;
-import com.limechain.grandpa.state.GrandpaSetState;
 import com.limechain.network.ConnectionManager;
 import com.limechain.network.protocol.blockannounce.messages.BlockAnnounceHandshakeBuilder;
 import com.limechain.network.protocol.grandpa.messages.GrandpaMessageType;
@@ -46,14 +45,14 @@ public class GrandpaEngine {
 
     protected ConnectionManager connectionManager;
     protected BlockAnnounceHandshakeBuilder handshakeBuilder;
-    protected GrandpaSetState grandpaSetState;
+    protected GrandpaMessageHandler grandpaMessageHandler;
 
     public GrandpaEngine() {
         warpSyncState = AppBean.getBean(WarpSyncState.class);
 
         connectionManager = ConnectionManager.getInstance();
         handshakeBuilder = new BlockAnnounceHandshakeBuilder();
-        grandpaSetState = AppBean.getBean(GrandpaSetState.class);
+        grandpaMessageHandler = AppBean.getBean(GrandpaMessageHandler.class);
     }
 
     /**
@@ -150,7 +149,7 @@ public class GrandpaEngine {
         new Thread(() -> warpSyncState.syncNeighbourMessage(neighbourMessage, peerId)).start();
 
         if (AbstractState.isActiveAuthority() && connectionManager.checkIfPeerIsAuthorNode(peerId)) {
-            grandpaSetState.initiateAndSendCatchUpRequest(neighbourMessage, peerId);
+            grandpaMessageHandler.initiateAndSendCatchUpRequest(neighbourMessage, peerId);
         }
     }
 
@@ -158,7 +157,7 @@ public class GrandpaEngine {
         ScaleCodecReader reader = new ScaleCodecReader(message);
         VoteMessage voteMessage = reader.read(VoteMessageScaleReader.getInstance());
         //Maybe we need to add possible roundNumber check
-        grandpaSetState.handleVoteMessage(voteMessage);
+        grandpaMessageHandler.handleVoteMessage(voteMessage);
         //todo: handle vote message (authoring node responsibility?)
         log.log(Level.INFO, "Received vote message from Peer " + peerId + "\n" + voteMessage);
     }
@@ -175,7 +174,7 @@ public class GrandpaEngine {
         log.log(Level.INFO, "Received catch up request message from Peer " + peerId + "\n" + catchUpReqMessage);
 
         if (AbstractState.isActiveAuthority() && connectionManager.checkIfPeerIsAuthorNode(peerId)) {
-            grandpaSetState.initiateAndSendCatchUpResponse(peerId, catchUpReqMessage, connectionManager::getPeerIds);
+            grandpaMessageHandler.initiateAndSendCatchUpResponse(peerId, catchUpReqMessage, connectionManager::getPeerIds);
         }
     }
 
@@ -255,7 +254,7 @@ public class GrandpaEngine {
     /**
      * Send our GRANDPA vote message from {@link GrandpaService} on a given <b>responder</b> stream.
      *
-     * @param stream               <b>responder</b> stream to write the message to
+     * @param stream             <b>responder</b> stream to write the message to
      * @param encodedVoteMessage scale encoded VoteMessage object
      */
     public void writeVoteMessage(Stream stream, byte[] encodedVoteMessage) {
